corrected: moved the sendResetPasswordEmail method to ResetPasswordService
