+++ social-network-api/src/main/java/com/facebook/service/EmailHandlerService.java
@@ -8,14 +8,20 @@ import org.simplejavamail.api.mailer.config.TransportStrategy;
 import org.simplejavamail.email.EmailBuilder;
 import org.simplejavamail.mailer.MailerBuilder;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.retry.annotation.Backoff;
+import org.springframework.retry.annotation.Recover;
+import org.springframework.retry.annotation.Retryable;
 import org.springframework.stereotype.Service;
 
 @Log4j2
 @Service
-public class EmailService {
+public class EmailHandlerService {
+
+    private final static int MAX_ATTEMPTS = 3;
 
     @Value("${email.host}")
     private String host;
+
     @Value("${email.port}")
     private int port;
 
@@ -25,6 +31,7 @@ public class EmailService {
     @Value("${email.password}")
     private String password;
 
+    @Retryable(retryFor = MailException.class, maxAttempts = MAX_ATTEMPTS, backoff = @Backoff(delay = 10000))
     public void sendEmail(String to, String subject, String messageContent) throws Exception {
         Email email = EmailBuilder.startingBlank()
                 .from(username)
@@ -32,26 +39,22 @@ public class EmailService {
                 .withSubject(subject)
                 .withHTMLText(messageContent)
                 .buildEmail();
+
         try (Mailer mailer = MailerBuilder
                 .withSMTPServer(host, port, username, password)
                 .withTransportStrategy(TransportStrategy.SMTPS)  // SMTPS = SMTP + SSL
                 .buildMailer()) {
 
             mailer.sendMail(email);
-        } catch (MailException e) {
-            log.error("Ошибка при отправке письма: ", e);
-            // Для повторної спроби надсилання листа:
-            // retrySendEmail(to, subject, messageContent);
         }
     }
 
-    public void sendResetPasswordEmail(String email, String token, String url) throws Exception {
-        String resetPasswordLetterSubject = "Reset password";
-        String resetPasswordLetterContent ="<p>Click the link below to reset your password:<br>"
-                +"<a href=%s>Reset password</a>"
-                +"<br>This link is valid for 15 minutes.<br>"
-                +"If you didn't request password change just ignore this letter.</div>";
-        sendEmail(email, resetPasswordLetterSubject,
-                String.format(resetPasswordLetterContent, url +"/"+ token +"?em=" + email));
+    // Метод автоматично викликається після того,
+    // як досягнуто максимальної кількості спроб надсилання листа.
+    @Recover
+    public void handleMailException(MailException e, String to, String subject, String messageContent) {
+        log.error("Не вдалося надіслати листа після {} спроб: ", MAX_ATTEMPTS, e);
+
     }
+
 }
