+++ social-network-api/src/main/java/com/facebook/service/ResetPasswordService.java
@@ -5,8 +5,8 @@ import com.facebook.dto.appuser.UserNewPasswordRequest;
 import com.facebook.exception.InvalidTokenException;
 import com.facebook.exception.UserNotFoundException;
 import com.facebook.model.AppUser;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.Optional;
@@ -14,18 +14,11 @@ import java.util.UUID;
 
 @Log4j2
 @Service
+@RequiredArgsConstructor
 public class ResetPasswordService {
     private final CacheStore<String> resetPasswordTokenCache;
-    private final EmailService emailService;
+    private final EmailHandlerService emailHandlerService;
     private final AppUserService appUserService;
-
-
-    @Autowired
-    public ResetPasswordService(CacheStore<String> resetPasswordTokenCache, EmailService emailService, AppUserService appUserService) {
-        this.resetPasswordTokenCache = resetPasswordTokenCache;
-        this.emailService = emailService;
-        this.appUserService = appUserService;
-    }
 
     String generateToken(){
         return UUID.randomUUID().toString();
@@ -47,7 +40,7 @@ public class ResetPasswordService {
         try{
             String resetPasswordToken = createAndAddResetPasswordToken(email);
             log.info("token " + resetPasswordToken);
-            emailService.sendResetPasswordEmail(email, resetPasswordToken, url);
+            sendResetPasswordEmail(email, resetPasswordToken, url);
         } catch (Exception e){
             log.error("Error sending reset password email", e);
         }
@@ -58,5 +51,16 @@ public class ResetPasswordService {
         appUserService.updatePassword(user.getEmail(), user.getNewPassword());
         resetPasswordTokenCache.removeToken(user.getEmail());
     }
+
+    public void sendResetPasswordEmail(String email, String token, String url) throws Exception {
+        String resetPasswordLetterSubject = "Reset password";
+        String resetPasswordLetterContent ="<p>Click the link below to reset your password:<br>"
+                +"<a href=%s>Reset password</a>"
+                +"<br>This link is valid for 15 minutes.<br>"
+                +"If you didn't request password change just ignore this letter.</div>";
+        emailHandlerService.sendEmail(email, resetPasswordLetterSubject,
+                String.format(resetPasswordLetterContent, url +"/"+ token +"?em=" + email));
+    }
+
 }
